check for string true rather than existence in setup-python-keyring (#41)
We saw builds breaking because `python -m pip install
"keyrings-chainguard-libraries"` was executing on custom runners which
don't have the `python` executable installed. The intent seems to be
that this step only runs when `setup-python-keyring` is overriden to
`true`, and not by default, but the current implementation will always
run it (the `if` statement as currently written evaluates to `true`
because the variable exists). This fixes the `if` statement to only
evaluate to `true` if `setup-python-keyring` has been overridden to
`true`.